+++ include/metalang99/list.h
@@ -510,7 +510,7 @@
  * ML99_listEq(v(ML99_natEq), ML99_list(v(1, 2, 3)), ML99_list(v(1, 2, 3)))
  * @endcode
  */
-#define ML99_listEq(compare, list, other) ML99_call(ML99_listEq, compare, list, other)
+#define ML99_listEq(cmp, list, other) ML99_call(ML99_listEq, cmp, list, other)
 
 /**
  * Checks whether @p item resides in @p list.
@@ -528,7 +528,7 @@
  * ML99_listContains(v(ML99_natEq), v(456), ML99_list(v(1, 2, 3)))
  * @endcode
  */
-#define ML99_listContains(compare, item, list) ML99_call(ML99_listContains, compare, item, list)
+#define ML99_listContains(cmp, item, list) ML99_call(ML99_listContains, cmp, item, list)
 
 /**
  * Extracts the prefix of @p list of the length @p n. If @p n is greater than the length of @p list,
@@ -591,7 +591,7 @@
 #define ML99_listDropWhile(f, list) ML99_call(ML99_listDropWhile, f, list)
 
 /**
- * Computes a list of corresponding pairs from @p list and @p other.
+ * Computes a list of corresponding pairs (as tuples) from @p list and @p other.
  *
  * # Examples
  *
@@ -599,22 +599,22 @@
  * #include <metalang99/list.h>
  *
  * // (1, 4), (2, 5), (3, 6)
- * E_zip(ML99_list(v(1, 2, 3)), ML99_list(v(4, 5, 6)))
+ * ML99_listZip(ML99_list(v(1, 2, 3)), ML99_list(v(4, 5, 6)))
  * @endcode
  */
 #define ML99_listZip(list, other) ML99_call(ML99_listZip, list, other)
 
 /**
- * Transforms a list of pairs into a list of the first components and a list of the second
- * components.
+ * Transforms a list of pairs (as tuples) into a tuple of a list of the first components and a list
+ * of the second components.
  *
  * # Examples
  *
  * @code
  * #include <metalang99/list.h>
  *
- * // ML99_list(ML99_list(v(1, 2, 3)), ML99_list(v(4, 5, 6)))
- * ML99_listUnzip(ML99_list(ML99_list(v(1, 4)), ML99_list(v(2, 5)), ML99_list(v(3, 6))))
+ * // ML99_tuple(ML99_list(v(1, 2, 3)), ML99_list(v(4, 5, 6)))
+ * ML99_listUnzip(ML99_list(ML99_tuple(v(1, 4)), ML99_tuple(v(2, 5)), ML99_tuple(v(3, 6))))
  * @endcode
  */
 #define ML99_listUnzip(list) ML99_call(ML99_listUnzip, list)
@@ -637,8 +637,8 @@
 #define ML99_listReplicate(n, item) ML99_call(ML99_listReplicate, n, item)
 
 /**
- * Returns a pair of list: those items of @p list the do and do not satisfy the predicate @p f,
- * respectively.
+ * Returns a pair (as tuple) of lists: those items of @p list the do and do not satisfy the
+ * predicate @p f, respectively.
  *
  * # Examples
  *
@@ -646,7 +646,7 @@
  * #include <metalang99/list.h>
  * #include <metalang99/nat.h>
  *
- * // ML99_list(ML99_list(v(4, 7)) ML99_list(v(11, 12, 13)))
+ * // ML99_tuple(ML99_list(v(4, 7)), ML99_list(v(11, 12, 13)))
  * ML99_listPartition(ML99_appl(v(ML99_greater), v(10)), ML99_list(v(11, 4, 12, 13, 7)))
  * @endcode
  */
@@ -683,7 +683,7 @@
 #ifndef DOXYGEN_IGNORE
 
 #define ML99_cons_IMPL(x, xs) v(ML99_CONS(x, xs))
-#define ML99_nil_IMPL(_)      v(ML99_NIL())
+#define ML99_nil_IMPL()       v(ML99_NIL())
 
 // ML99_listHead_IMPL {
 #define ML99_listHead_IMPL(list)             ML99_match_IMPL(list, ML99_PRIV_listHead_)
@@ -701,25 +701,14 @@
 #define ML99_listLast_IMPL(list)       ML99_match_IMPL(list, ML99_PRIV_listLast_)
 #define ML99_PRIV_listLast_nil_IMPL(_) ML99_PRIV_EMPTY_LIST_ERROR(listLast)
 #define ML99_PRIV_listLast_cons_IMPL(x, xs)                                                        \
-    ML99_call(                                                                                     \
-        ML99_callUneval(                                                                           \
-            ML99_if,                                                                               \
-            ML99_IS_NIL(xs),                                                                       \
-            ML99_PRIV_listLastDone,                                                                \
-            ML99_PRIV_listLastProgress),                                                           \
-        v(x, xs))
-#define ML99_PRIV_listLastDone_IMPL(x, _xs)     v(x)
-#define ML99_PRIV_listLastProgress_IMPL(_x, xs) ML99_listLast_IMPL(xs)
+    ML99_IF(ML99_IS_NIL(xs), v(x), ML99_listLast_IMPL(xs))                                         \
 // }
 
 // ML99_listInit_IMPL {
 #define ML99_listInit_IMPL(list)       ML99_match_IMPL(list, ML99_PRIV_listInit_)
 #define ML99_PRIV_listInit_nil_IMPL(_) ML99_PRIV_EMPTY_LIST_ERROR(listInit)
 #define ML99_PRIV_listInit_cons_IMPL(x, xs)                                                        \
-    ML99_call(                                                                                     \
-        ML99_callUneval(ML99_if, ML99_IS_NIL(xs), ML99_PRIV_constNil, ML99_PRIV_listInitProgress), \
-        v(x, xs))
-#define ML99_PRIV_listInitProgress_IMPL(x, xs) ML99_cons(v(x), ML99_listInit_IMPL(xs))
+    ML99_IF(ML99_IS_NIL(xs), v(ML99_NIL()), ML99_cons(v(x), ML99_listInit_IMPL(xs)))
 // }
 
 // ML99_list_IMPL {
@@ -777,23 +766,20 @@
 #define ML99_listAppendItem_IMPL(item, list) ML99_listAppend_IMPL(list, ML99_CONS(item, ML99_NIL()))
 // }
 
-// ML99_isCons_IMPL {
+// ML99_isCons_IMPL, ML99_isNil_IMPL {
 #define ML99_isCons_IMPL(list) v(ML99_IS_CONS(list))
-// }
-
-// ML99_isNil_IMPL {
-#define ML99_isNil_IMPL(list) v(ML99_IS_NIL(list))
+#define ML99_isNil_IMPL(list)  v(ML99_IS_NIL(list))
 // }
 
 // ML99_listUnwrap_IMPL {
 #define ML99_listUnwrap_IMPL(list)            ML99_match_IMPL(list, ML99_PRIV_listUnwrap_)
-#define ML99_PRIV_listUnwrap_nil_IMPL(_)      ML99_empty()
+#define ML99_PRIV_listUnwrap_nil_IMPL(_)      v(ML99_EMPTY())
 #define ML99_PRIV_listUnwrap_cons_IMPL(x, xs) ML99_TERMS(v(x), ML99_listUnwrap_IMPL(xs))
 // }
 
 // ML99_listReverse_IMPL {
 #define ML99_listReverse_IMPL(list)            ML99_match_IMPL(list, ML99_PRIV_listReverse_)
-#define ML99_PRIV_listReverse_nil_IMPL(_)      ML99_nil()
+#define ML99_PRIV_listReverse_nil_IMPL(_)      v(ML99_NIL())
 #define ML99_PRIV_listReverse_cons_IMPL(x, xs) ML99_listAppendItem(v(x), ML99_listReverse_IMPL(xs))
 // }
 
@@ -801,10 +787,7 @@
 #define ML99_listGet_IMPL(i, list)       ML99_matchWithArgs_IMPL(list, ML99_PRIV_listGet_, i)
 #define ML99_PRIV_listGet_nil_IMPL(_, i) ML99_PRIV_EMPTY_LIST_ERROR(ML99_get)
 #define ML99_PRIV_listGet_cons_IMPL(x, xs, i)                                                      \
-    ML99_IF(ML99_NAT_EQ(i, 0), ML99_PRIV_listGetDone, ML99_PRIV_listGetProgress)                   \
-    (x, ML99_DEC(i), xs)
-#define ML99_PRIV_listGetDone(x, _i, _xs)   v(x)
-#define ML99_PRIV_listGetProgress(x, i, xs) ML99_listGet_IMPL(i, xs)
+    ML99_IF(ML99_NAT_EQ(i, 0), v(x), ML99_listGet_IMPL(ML99_DEC(i), xs))
 // }
 
 // ML99_listFoldr_IMPL {
@@ -832,7 +815,7 @@
 // ML99_listIntersperse_IMPL {
 #define ML99_listIntersperse_IMPL(item, list)                                                      \
     ML99_matchWithArgs_IMPL(list, ML99_PRIV_listIntersperse_, item)
-#define ML99_PRIV_listIntersperse_nil_IMPL(_, _item) ML99_nil()
+#define ML99_PRIV_listIntersperse_nil_IMPL(...) v(ML99_NIL())
 #define ML99_PRIV_listIntersperse_cons_IMPL(x, xs, item)                                           \
     ML99_cons(v(x), ML99_listPrependToAll_IMPL(item, xs))
 // }
@@ -840,14 +823,14 @@
 // ML99_listPrependToAll_IMPL {
 #define ML99_listPrependToAll_IMPL(item, list)                                                     \
     ML99_matchWithArgs_IMPL(list, ML99_PRIV_listPrependToAll_, item)
-#define ML99_PRIV_listPrependToAll_nil_IMPL(_, _item) ML99_nil()
+#define ML99_PRIV_listPrependToAll_nil_IMPL(...) v(ML99_NIL())
 #define ML99_PRIV_listPrependToAll_cons_IMPL(x, xs, item)                                          \
     ML99_cons(v(item), ML99_cons(v(x), ML99_listPrependToAll_IMPL(item, xs)))
 // }
 
 // ML99_listMap_IMPL {
-#define ML99_listMap_IMPL(f, list)        ML99_matchWithArgs_IMPL(list, ML99_PRIV_listMap_, f)
-#define ML99_PRIV_listMap_nil_IMPL(_, _f) ML99_nil()
+#define ML99_listMap_IMPL(f, list)      ML99_matchWithArgs_IMPL(list, ML99_PRIV_listMap_, f)
+#define ML99_PRIV_listMap_nil_IMPL(...) v(ML99_NIL())
 #define ML99_PRIV_listMap_cons_IMPL(x, xs, f)                                                      \
     ML99_cons(ML99_appl_IMPL(f, x), ML99_listMap_IMPL(f, xs))
 // }
@@ -856,7 +839,7 @@
 #define ML99_listMapI_IMPL(f, list) ML99_PRIV_listMapIAux_IMPL(f, list, 0)
 #define ML99_PRIV_listMapIAux_IMPL(f, list, i)                                                     \
     ML99_matchWithArgs_IMPL(list, ML99_PRIV_listMapI_, f, i)
-#define ML99_PRIV_listMapI_nil_IMPL(_, _f, _i) ML99_nil()
+#define ML99_PRIV_listMapI_nil_IMPL(...) v(ML99_NIL())
 #define ML99_PRIV_listMapI_cons_IMPL(x, xs, f, i)                                                  \
     ML99_cons(ML99_appl2_IMPL(f, x, i), ML99_PRIV_listMapIAux_IMPL(f, xs, ML99_INC(i)))
 // }
@@ -864,7 +847,7 @@
 // ML99_listMapInPlace_IMPL {
 #define ML99_listMapInPlace_IMPL(f, list)                                                          \
     ML99_matchWithArgs_IMPL(list, ML99_PRIV_listMapInPlace_, f)
-#define ML99_PRIV_listMapInPlace_nil_IMPL(_, _f) ML99_empty()
+#define ML99_PRIV_listMapInPlace_nil_IMPL(...) v(ML99_EMPTY())
 #define ML99_PRIV_listMapInPlace_cons_IMPL(x, xs, f)                                               \
     ML99_TERMS(ML99_appl_IMPL(f, x), ML99_listMapInPlace_IMPL(f, xs))
 // }
@@ -873,7 +856,7 @@
 #define ML99_listMapInPlaceI_IMPL(f, list) ML99_PRIV_listMapInPlaceIAux_IMPL(f, list, 0)
 #define ML99_PRIV_listMapInPlaceIAux_IMPL(f, list, i)                                              \
     ML99_matchWithArgs_IMPL(list, ML99_PRIV_listMapInPlaceI_, f, i)
-#define ML99_PRIV_listMapInPlaceI_nil_IMPL(_, _f, _i) ML99_empty()
+#define ML99_PRIV_listMapInPlaceI_nil_IMPL(...) v(ML99_EMPTY())
 #define ML99_PRIV_listMapInPlaceI_cons_IMPL(x, xs, f, i)                                           \
     ML99_TERMS(ML99_appl2_IMPL(f, x, i), ML99_PRIV_listMapInPlaceIAux_IMPL(f, xs, ML99_INC(i)))
 // }
@@ -895,8 +878,8 @@
 // }
 
 // ML99_listFilter_IMPL {
-#define ML99_listFilter_IMPL(f, list)        ML99_matchWithArgs_IMPL(list, ML99_PRIV_listFilter_, f)
-#define ML99_PRIV_listFilter_nil_IMPL(_, _f) ML99_nil()
+#define ML99_listFilter_IMPL(f, list)      ML99_matchWithArgs_IMPL(list, ML99_PRIV_listFilter_, f)
+#define ML99_PRIV_listFilter_nil_IMPL(...) v(ML99_NIL())
 #define ML99_PRIV_listFilter_cons_IMPL(x, xs, f)                                                   \
     ML99_call(                                                                                     \
         ML99_call(ML99_if, ML99_appl_IMPL(f, x), v(ML99_cons, ML99_PRIV_listFilterRest)),          \
@@ -906,49 +889,46 @@
 // }
 
 // ML99_listEq_IMPL {
-#define ML99_listEq_IMPL(compare, list, other)                                                     \
-    ML99_matchWithArgs_IMPL(list, ML99_PRIV_listEq_, other, compare)
+#define ML99_listEq_IMPL(cmp, list, other)                                                         \
+    ML99_matchWithArgs_IMPL(list, ML99_PRIV_listEq_, other, cmp)
 
-#define ML99_PRIV_listEq_nil_IMPL(_, other, _compare) v(ML99_IS_NIL(other))
-#define ML99_PRIV_listEq_cons_IMPL(x, xs, other, compare)                                          \
-    ML99_matchWithArgs_IMPL(other, ML99_PRIV_listEq_cons_, x, xs, compare)
+#define ML99_PRIV_listEq_nil_IMPL(_, other, _cmp) v(ML99_IS_NIL(other))
+#define ML99_PRIV_listEq_cons_IMPL(x, xs, other, cmp)                                              \
+    ML99_matchWithArgs_IMPL(other, ML99_PRIV_listEq_cons_, x, xs, cmp)
 
-#define ML99_PRIV_listEq_cons_nil_IMPL(_, _x, _xs, _compare) ML99_false
-#define ML99_PRIV_listEq_cons_cons_IMPL(other_x, other_xs, x, xs, compare)                         \
+#define ML99_PRIV_listEq_cons_nil_IMPL(...) ML99_false
+#define ML99_PRIV_listEq_cons_cons_IMPL(other_x, other_xs, x, xs, cmp)                             \
     ML99_call(                                                                                     \
         ML99_call(                                                                                 \
             ML99_if,                                                                               \
-            ML99_appl2_IMPL(compare, x, other_x),                                                  \
+            ML99_appl2_IMPL(cmp, x, other_x),                                                      \
             v(ML99_listEq, ML99_PRIV_constFalse)),                                                 \
-        v(compare, xs, other_xs))
+        v(cmp, xs, other_xs))
 // }
 
 // ML99_listContains_IMPL {
-#define ML99_listContains_IMPL(compare, item, list)                                                \
-    ML99_matchWithArgs_IMPL(list, ML99_PRIV_listContains_, item, compare)
-#define ML99_PRIV_listContains_nil_IMPL(_, _item, _compare) ML99_false
-#define ML99_PRIV_listContains_cons_IMPL(x, xs, item, compare)                                     \
+#define ML99_listContains_IMPL(cmp, item, list)                                                    \
+    ML99_matchWithArgs_IMPL(list, ML99_PRIV_listContains_, item, cmp)
+#define ML99_PRIV_listContains_nil_IMPL(...) ML99_false
+#define ML99_PRIV_listContains_cons_IMPL(x, xs, item, cmp)                                         \
     ML99_call(                                                                                     \
         ML99_call(                                                                                 \
             ML99_if,                                                                               \
-            ML99_appl2_IMPL(compare, x, item),                                                     \
+            ML99_appl2_IMPL(cmp, x, item),                                                         \
             v(ML99_PRIV_constTrue, ML99_listContains)),                                            \
-        v(compare, item, xs))
+        v(cmp, item, xs))
 // }
 
 // ML99_listTake_IMPL {
-#define ML99_listTake_IMPL(n, list)        ML99_matchWithArgs_IMPL(list, ML99_PRIV_listTake_, n)
-#define ML99_PRIV_listTake_nil_IMPL(_, _i) ML99_nil()
+#define ML99_listTake_IMPL(n, list)      ML99_matchWithArgs_IMPL(list, ML99_PRIV_listTake_, n)
+#define ML99_PRIV_listTake_nil_IMPL(...) v(ML99_NIL())
 #define ML99_PRIV_listTake_cons_IMPL(x, xs, i)                                                     \
-    ML99_IF(ML99_NAT_EQ(i, 0), ML99_PRIV_listTakeDone, ML99_PRIV_listTakeProgress)                 \
-    (x, xs, i)
-#define ML99_PRIV_listTakeDone(_x, _xs, _i)  ML99_nil()
-#define ML99_PRIV_listTakeProgress(x, xs, i) ML99_cons(v(x), ML99_listTake_IMPL(ML99_DEC(i), xs))
+    ML99_IF(ML99_NAT_EQ(i, 0), v(ML99_NIL()), ML99_cons(v(x), ML99_listTake_IMPL(ML99_DEC(i), xs)))
 // }
 
 // ML99_listTakeWhile_IMPL {
-#define ML99_listTakeWhile_IMPL(f, list)        ML99_matchWithArgs_IMPL(list, ML99_PRIV_listTakeWhile_, f)
-#define ML99_PRIV_listTakeWhile_nil_IMPL(_, _f) ML99_nil()
+#define ML99_listTakeWhile_IMPL(f, list)      ML99_matchWithArgs_IMPL(list, ML99_PRIV_listTakeWhile_, f)
+#define ML99_PRIV_listTakeWhile_nil_IMPL(...) v(ML99_NIL())
 #define ML99_PRIV_listTakeWhile_cons_IMPL(x, xs, f)                                                \
     ML99_call(                                                                                     \
         ML99_call(                                                                                 \
@@ -961,18 +941,15 @@
 // }
 
 // ML99_listDrop_IMPL {
-#define ML99_listDrop_IMPL(n, list)        ML99_matchWithArgs_IMPL(list, ML99_PRIV_listDrop_, n)
-#define ML99_PRIV_listDrop_nil_IMPL(_, _i) ML99_nil()
+#define ML99_listDrop_IMPL(n, list)      ML99_matchWithArgs_IMPL(list, ML99_PRIV_listDrop_, n)
+#define ML99_PRIV_listDrop_nil_IMPL(...) v(ML99_NIL())
 #define ML99_PRIV_listDrop_cons_IMPL(x, xs, i)                                                     \
-    ML99_IF(ML99_NAT_EQ(i, 0), ML99_PRIV_listDropDone, ML99_PRIV_listDropProgress)                 \
-    (x, xs, i)
-#define ML99_PRIV_listDropDone(x, xs, _i)     ML99_call(ML99_cons, v(x, xs))
-#define ML99_PRIV_listDropProgress(_x, xs, i) ML99_listDrop_IMPL(ML99_DEC(i), xs)
+    ML99_IF(ML99_NAT_EQ(i, 0), v(ML99_CONS(x, xs)), ML99_listDrop_IMPL(ML99_DEC(i), xs))
 // }
 
 // ML99_listDropWhile_IMPL {
-#define ML99_listDropWhile_IMPL(f, list)        ML99_matchWithArgs_IMPL(list, ML99_PRIV_listDropWhile_, f)
-#define ML99_PRIV_listDropWhile_nil_IMPL(_, _f) ML99_nil()
+#define ML99_listDropWhile_IMPL(f, list)      ML99_matchWithArgs_IMPL(list, ML99_PRIV_listDropWhile_, f)
+#define ML99_PRIV_listDropWhile_nil_IMPL(...) v(ML99_NIL())
 #define ML99_PRIV_listDropWhile_cons_IMPL(x, xs, f)                                                \
     ML99_call(                                                                                     \
         ML99_call(                                                                                 \
@@ -987,31 +964,33 @@
 // ML99_listZip_IMPL {
 #define ML99_listZip_IMPL(list, other) ML99_matchWithArgs_IMPL(list, ML99_PRIV_listZip_, other)
 
-#define ML99_PRIV_listZip_nil_IMPL(_, _other) ML99_nil()
+#define ML99_PRIV_listZip_nil_IMPL(...) v(ML99_NIL())
 #define ML99_PRIV_listZip_cons_IMPL(x, xs, other)                                                  \
     ML99_matchWithArgs_IMPL(other, ML99_PRIV_listZip_cons_, x, xs)
 
-#define ML99_PRIV_listZip_cons_nil_IMPL(_, _x, _xs) ML99_nil()
+#define ML99_PRIV_listZip_cons_nil_IMPL(...) v(ML99_NIL())
 #define ML99_PRIV_listZip_cons_cons_IMPL(other_x, other_xs, x, xs)                                 \
-    ML99_cons(ML99_list_IMPL(x, other_x), ML99_listZip_IMPL(xs, other_xs))
+    ML99_cons(v(ML99_TUPLE(x, other_x)), ML99_listZip_IMPL(xs, other_xs))
 // }
 
 // ML99_listUnzip_IMPL {
 #define ML99_listUnzip_IMPL(list) ML99_match_IMPL(list, ML99_PRIV_listUnzip_)
 
-#define ML99_PRIV_listUnzip_nil_IMPL(_) ML99_list_IMPL(ML99_NIL(), ML99_NIL())
+#define ML99_PRIV_listUnzip_nil_IMPL(_) v(ML99_TUPLE(ML99_NIL(), ML99_NIL()))
 #define ML99_PRIV_listUnzip_cons_IMPL(x, xs)                                                       \
     ML99_call(ML99_PRIV_listUnzipProgress, v(x), ML99_listUnzip_IMPL(xs))
+
 #define ML99_PRIV_listUnzipProgress_IMPL(x, rest)                                                  \
-    ML99_list(ML99_PRIV_listUnzipAdd_IMPL(x, rest, 0), ML99_PRIV_listUnzipAdd_IMPL(x, rest, 1))
-#define ML99_PRIV_listUnzipAdd_IMPL(x, rest, i)                                                    \
-    ML99_cons(ML99_listGet_IMPL(i, x), ML99_listGet_IMPL(i, rest))
+    v(ML99_TUPLE(ML99_PRIV_LIST_UNZIP_EXTEND(x, rest, 0), ML99_PRIV_LIST_UNZIP_EXTEND(x, rest, 1)))
+
+#define ML99_PRIV_LIST_UNZIP_EXTEND(x, rest, i)                                                    \
+    ML99_CONS(ML99_TUPLE_GET(i)(x), ML99_TUPLE_GET(i)(rest))
 // }
 
 // ML99_listReplicate_IMPL {
 #define ML99_listReplicate_IMPL(n, item)                                                           \
     ML99_natMatchWithArgs_IMPL(n, ML99_PRIV_listReplicate_, item)
-#define ML99_PRIV_listReplicate_Z_IMPL(item)    ML99_nil()
+#define ML99_PRIV_listReplicate_Z_IMPL(...)     v(ML99_NIL())
 #define ML99_PRIV_listReplicate_S_IMPL(n, item) ML99_cons(v(item), ML99_listReplicate_IMPL(n, item))
 // }
 
@@ -1019,7 +998,7 @@
 #define ML99_listPartition_IMPL(f, list)                                                           \
     ML99_listFoldr(                                                                                \
         ML99_appl_IMPL(ML99_PRIV_listPartitionAux, f),                                             \
-        ML99_list_IMPL(ML99_NIL(), ML99_NIL()),                                                    \
+        v(ML99_TUPLE(ML99_NIL(), ML99_NIL())),                                                     \
         v(list))
 
 #define ML99_PRIV_listPartitionAux_IMPL(f, x, acc)                                                 \
@@ -1027,13 +1006,11 @@
         ML99_call(                                                                                 \
             ML99_if,                                                                               \
             ML99_appl_IMPL(f, x),                                                                  \
-            v(ML99_PRIV_listPartitionAuxExtendFst, ML99_PRIV_listPartitionAuxExtendSnd)),          \
-        v(x),                                                                                      \
-        ML99_listGet_IMPL(0, acc),                                                                 \
-        ML99_listGet_IMPL(1, acc))
+            v(ML99_PRIV_listPartitionExtendFst, ML99_PRIV_listPartitionExtendSnd)),                \
+        v(x, ML99_TUPLE_GET(0)(acc), ML99_TUPLE_GET(1)(acc)))
 
-#define ML99_PRIV_listPartitionAuxExtendFst_IMPL(x, fst, snd) ML99_list_IMPL(ML99_CONS(x, fst), snd)
-#define ML99_PRIV_listPartitionAuxExtendSnd_IMPL(x, fst, snd) ML99_list_IMPL(fst, ML99_CONS(x, snd))
+#define ML99_PRIV_listPartitionExtendFst_IMPL(x, fst, snd) v(ML99_TUPLE(ML99_CONS(x, fst), snd))
+#define ML99_PRIV_listPartitionExtendSnd_IMPL(x, fst, snd) v(ML99_TUPLE(fst, ML99_CONS(x, snd)))
 // }
 
 // ML99_listAppl_IMPL {
@@ -1044,12 +1021,12 @@
 #define ML99_listUnwrapCommaSep_IMPL(list)                                                         \
     ML99_IF(                                                                                       \
         ML99_IS_NIL(list),                                                                         \
-        ML99_empty(),                                                                              \
+        v(ML99_EMPTY()),                                                                           \
         ML99_variadicsTail(ML99_PRIV_listUnwrapCommaSepAux_IMPL(list)))
 
 #define ML99_PRIV_listUnwrapCommaSepAux_IMPL(xs)                                                   \
     ML99_match_IMPL(xs, ML99_PRIV_listUnwrapCommaSepAux_)
-#define ML99_PRIV_listUnwrapCommaSepAux_nil_IMPL(_) ML99_empty()
+#define ML99_PRIV_listUnwrapCommaSepAux_nil_IMPL(_) v(ML99_EMPTY())
 #define ML99_PRIV_listUnwrapCommaSepAux_cons_IMPL(x, xs)                                           \
     ML99_TERMS(v(, x), ML99_PRIV_listUnwrapCommaSepAux_IMPL(xs))
 // }
@@ -1062,7 +1039,7 @@
 #define ML99_PRIV_IS_NIL_nil()  ML99_TRUE
 #define ML99_PRIV_IS_NIL_cons() ML99_FALSE
 
-#define ML99_PRIV_constNil_IMPL(...) ML99_nil()
+#define ML99_PRIV_constNil_IMPL(...) v(ML99_NIL())
 
 // Arity specifiers {
 #define ML99_cons_ARITY               2
@@ -1106,15 +1083,7 @@
 #define ML99_listPartition_ARITY      2
 #define ML99_listAppl_ARITY           2
 
-#define ML99_PRIV_listInitProgress_ARITY      2
-#define ML99_PRIV_listDone_ARITY              1
-#define ML99_PRIV_listProgress_ARITY          2
-#define ML99_PRIV_listMapIProvideIdx_ARITY    3
-#define ML99_PRIV_listTakeProgress_ARITY      3
-#define ML99_PRIV_listTakeWhileProgress_ARITY 3
-#define ML99_PRIV_listDropProgress_ARITY      2
-#define ML99_PRIV_listDropWhileProgress_ARITY 2
-#define ML99_PRIV_listPartitionAux_ARITY      3
+#define ML99_PRIV_listPartitionAux_ARITY 3
 // }
 
 #endif // DOXYGEN_IGNORE